Fix: duplicate sub-headers appearing in sidebar
Hovering on a page preview link adds duplicate sub-headers to the
sidebar. This change hides the loaded preview content from Docsify,
which was scanning this content for sub-headers.

This reverts commit 56fe7969a70daade17b6dd6a5c2a3153998ac352, which was
the previous fix for this issue. Further testing indicated it did not
work.

Note: as a workaround, this fix loads a duplicate version of the
Markdown parsing library, rather than re-use the Docsify instance. The
Docsify instance is stateful and has numerous side effects, one of
which causes this issue.