Fix: not displaying nested sections in popup
We are not correctly displaying previews for deep linked sections. This
is because the output HTML was not being correctly sanitised.

Note: as a workaround, this fix loads a duplicate version of the
HTML sanitising library already bundled with Docsify. This is because I
could not find a way to copy the markdown configuration properties, or
reliably import the HTML sanitising library from Docsify.